requirements.txt
Django==1.10.3
django-easy-select2==1.3.3
django-material==0.10.1
Pillow==3.4.2
